Change ContentfulComponent to better ComponentRenderer component, duh
