+++ packages/contentful-core/src/components/index.ts
@@ -1,0 +1,1 @@
+export * from './ComponentRenderer';
